Copy the device when using the Ghostscript API
The documentation of finddevice says:
> Creates a default instance of a device specified by name. The
> instance is created in global VM. If finddevice is called more
> than once with the same device name, it creates the default
> instance the first time, and returns the same instance thereafter.

As a result, multiple commands sent to the API use the same output
device with an ever increasing PageCount property. This manifests
when using '%d' in the output filename, for example when LilyPond
creates one PNG file per page. See #5996 for details.
To solve this problem, create a fresh copy of the prototype device
before putdeviceprops and setdevice. This discards any modification
after the output has been written. However this leads to crashes
when creating multiple copies of the pdfwrite device. Luckily we
do not use '%d' when writing PDFs.