Fix URL generation for bug-list in Doc footer
IIUC this has been broken since
d79f0bb3738a3fffae211e8e8c550fad70cf55ec
The logic in process_html_files() checks,
whether 'mail_address' refers to an actual
mail address or to a web page. It did only
check for http://, not for https://, however.

Since the link to the bug-lilypond web interface
is hardcoded, there is no point in checking
whether it is an email address or an URL, thus
removed the outdated branching code.