Add PDF bookmark support & multi-level TOC outline ability.
- Support basic bookmarking pdfmark syntax (this can be
switched on by setting the new 'outline-bookmarks option,
which is disabled by default for now).

- Add an optional arg to \tocItem so that users may specify
a name or a hierarchical list of names; that information is
then stored in the 'label-alist-table output-def variable.

- That implies that \tocItem without a symbol now needs
to take an explicit markup; add convert-rule accordingly.

- Use that information to print the \table-of-contents
in a more structured way (by using the new 'toc:level
layout variable, by default pointing to 'tocIndentMarkup
which is also new and user-settable).

- Modify the \page-ref markup command to account for these
situations (and use the aforementioned alist).

- Add doc, Changes items and a new toc-structured.ly regtest.