Support C++11 range-based for loops over SCM lists
ly_scm_list_t<T> is a generalized wrapper around a SCM list.  It can be
used with any type for which to_scm(T) and from_scm<T>() are supported,
but will most commonly be used via two aliases:

    ly_scm_list       alias for ly_scm_list_t<SCM>
    ly_smob_list<T>   alias for ly_scm_list_t<T *>

Dereferencing an iterator automatically calls from_scm<T> ().

The primary design goal is to fit the conventions of the STL as closely
as possible, and where it is not possible, to make the differences
obvious enough to avoid misleading.