Avoid scm_append_x for property values
As of a456a59 ("Copy alist instead of deep copy on Grob clone")
property values are potentially shared between different grobs,
leading to unexpected outcomes with in-place modifications of property
values.

In this case, the construction of the 'labels property created a
cyclic list, leading to an infinite loop with unbounded memory usage.

Fixes #5999