texi2html: Simplify print_navigation
The previous code provided a very generic solution to style multiple
table cells. This was overly complicated for generating HTML code for
exactly two different sets of buttons. Rewrite the implementation in
function lilypond_print_navigation for the current configuration. The
new code also hides the 'Index' link if there is none, e.g. in CG.