texi2html: Fix variable mistake
This code path is taken if the node is not found in the map and
$filename is still unset. I'm not sure why this has worked until
now for the documentation, but this leads to empty filenames and
subsequent errors for my tiny example texi files. This has also
occurred during my experiments with texi2html-5.0, so I think this
is the right way to fix this.