Use Table of Contents generated by texi2html
Instead of generating our own TOC, just use the one provided by
texi2html. This only requires matching the entries of the current
page and its ancestors to add class="toc_current".

Afterwards, use CSS to hide the TOC levels that we do not want to
show by default:
 * for manuals, starting from the third level;
 * for web, only make the top level visible by default.
Entries are still shown if one of the siblings is on the path to
the current page or it is a direct child of the current page.

The only visible change is on the web page where second level TOCs
no longer have the entry from the top level as the first element.
In my opinion, that's an advantage because the lists were getting
too long in some cases anyway, see "Community".