Write output files atomically
Rather than opening the destination and writing to it, open a
temporary destination in the same directory as the destination, and
rename it into place after closing.

This protects against creating corrupted files, for example, in case of
interrupts to lilypond (eg. pressing ^C), and concurrent execution of
multiple lilypond processes.

PDF files are especially susceptible to this, as GS only finalized
them once the next output file is started.

File corruption is disruptive when using lilypond-book, as
lilypond-book only checks for file existence, but cannot check
well-formedness of a file that is already in the cache.