Fix empty PDFs with Ghostscript
GS uses a signal to initialize the pdfwrite device, see
http://git.ghostscript.com/?p=ghostpdl.git;a=commit;h=87a963337cdf7c7cd8406a1fc1c81e91c1728140

An updated comment in sending the signal states that "this may not
work if the initial device is not pdfwrite". This applies to the
new code path after commit 017927b4d6 ("Unify calling convention
for command-line and API GS use"), so trigger the signal ourselves.

Fixes #6001, after much debugging together with Han-Wen.