Doc: Allow building with texi2html-5.0
The result is similar to the version produced by version 1.82, but
there are minor differences: Section numbers are not terminated by
a dot anymore and the added prefix "Appendix" cannot be translated.
Furthermore, there are a few formatting differences and quite some
link texts do not match the ones generated by the old version.
In short: The main version continues to be 1.82 for now.