Use error expression arg of check_scheme_arg more often
The parser's check_scheme_arg function has an optional argument that
is only used when flagging an error to indicate the expression not
meeting the predicate.  It's basically there for use cases where
make_music_from_simple is being used for creating the original
expression, in order to keep the error message from becoming
confusingly dissociated with the originating expression.

This adds a few uses of this argument.