Add Ghostscript commands to debug message if `-dgs-api=#f`
If `-dgs-api=#f` is used, LilyPond outputs a command line
for invoking Ghostscript as a debug message.
However, Ghostscript commands written to a temporary file
were not in the debug message.
It was difficult to debug by the method
loading the PostScript file output by LilyPond into another Ghostscript
because the Ghostscript commands are not known.

This commit adds the Ghostscript commands to the debug message.
This makes debugging with PostScript files easier.