Inline lilypond-book rules into regtest dir
This removes rules from make/lilypond-book-rules. The rules are very
general, for example:

  $(outdir)/%.html: %.html

Removing them avoids ambiguous rules in in the
Documentation/GNUmakefile (which uses the lilypond-book rules) down
the line.

Also remove the lysdoc template. The lysdoc template is used for
constructing the regtest itself. As its functionality is not part
lilypond-book, we don't have to test it here.

Remove extractpdfmark support from the lilypond-book regtests. The
PDFs aren't shipped so we don't care about their size.