Build website from Gitlab CI.
This adds a dummy toplevel Make target, and a 'website' stage in the
pipeline.

This is intended to run either as a scheduled or manually triggered
job.  A cronjob on lilypond.org can check for updates and download the
resulting artifact as a zip file, unpacking it into the directory in
the WWW root.