Fix Ghostscript command for selecting a pdfwrite device
We used `setdevice` to select the Ghostscript `pdfwrite` device.
However, it does not set some defaults,
which increased the number of drawing operators for the output PDF.

This commit makes using `selectdevice` instead of `setdevice`.
`selectdevice` sets the defaults and is described
in the current Ghostscript document as a method to select the device.
This reduces the number of drawing operators for the output PDF.