Python cleanup (VI): Fix invalid escape sequences
There were quite a few instances of strings with non-escaped
backslashes, like '\mark'. Fortunately or not, the interpreter's
behavior in such cases is to turn the resulting invalid escape
sequence into a backslash followed by a character, which is why
they worked so far. Still, this is not great style  (since
while '\mark' is the string \mark, '\niente' is a newline followed
by 'iente'). Use autopep8 to change these into raw strings, like r'\mark'.

The Python Language Reference says:

   Unrecognized escape sequences produce a DeprecationWarning.
   In a future Python version they will be a SyntaxWarning and
   eventually a SyntaxError.

(Although they eventually decided to extend the deprecation period
so the warning is silent at the moment.)

The autopep8 output had to be slightly postprocessed manually, see
https://github.com/hhatto/autopep8/issues/558