Fix Ghostscript command for selecting png devices
We use `setdevice` to select the Ghostscript png devices.
However, it does not set some defaults.

We want to use `selectdevice` instead of `setdevice`.
But using `selectdevice` can't reset the page number
because `selectdevice` uses `finddevice` internally.

This commit adds to use an undocumented `.setdefaultscreen` procedure
which is used inside `selectdevice` to set the defaults.
It works in Ghostscript 9.52 but may not work
if the internal implementation of `selectdevice` is changed in the future.