Allow "make dist" for in-tree builds
The previous recipe tests for untracked files, which are common
occurence in in-tree builds. This has limited usefulness, because our
Gitlab-based CI already prevents forgetting to commit files.

Require a git repository for running "make dist"; the gitless mode
was broken anyway, as it does

       test -d $(top-src-dir)/.git && rm -f $(RELEASE_OUT_FILES)

Get rid of local-dist target. It is only used for out/RELEASE_COMMIT
in the toplevel directory. For simplicity, move RELEASE_COMMIT to the
toplevel directory.

Instead of truncating RELEASE_COMMIT to 100 lines, use the --no-patch
option, which is likely what was really wanted.

Tested:

    rm -rf before ; tar xfz /tmp/b/out/lilypond-2.21.4.tar.gz ; mv \
        lilypond-2.21.4 before
    make dist
    rm -rf after ; tar xfz out/lilypond-2.21.4.tar.gz ; mv lilypond-2.21.4 after
    diff -urN before after  # manual inspection

    tar vtfz /tmp/b/out/lilypond-2.21.4.tar.gz > before.txt
    tar vtfz out/lilypond-2.21.4.tar.gz > after.txt
    # edit timestamps in after.txt
    diff -u before.txt after.txt # manual inspection