Scheme-based implementation of the new grob DurationLine
It continues a rhythmic event (usually a `NoteHead`, `Rest`, or
`NoteColumn` grob) with a line.  Available line styles are: `beam`,
`line`, `dashed-line`, `dotted-line`, `zigzag`, `trill`, and `none`.

Syntax is `<note>\-`.

Usually it stops at the next rhythmical event (per default
not at skips), if the context "pauses", and at the end of a score.

Three regtests are provided as well as some documentation.
Three context-properties are added: `startAtSkip`, `startAtNoteColumn`,
and `endAtSkip`.