Reformat all Python files
This is the first part of an attempt to have a consistent Python coding
style in line with PEP 8. The code is reformatted using autopep8. Only
whitespace is affected in this run (with the minor exception of line-ending
semicolons removed); long lines are wrapped to fit into 79 characters. The
new code should be purely equivalent to the former one. More aggressive
changes autopep8 is capable of are done one at a time in further commits.

However, the autopep8 output needed a slight tweaking: the lilylib import
must come after gettext.install(...) when present.