web: Remove page with Acknowledgements
This list was initially added to thank mshparisnord.fr for automated
build tests. As far as I can tell, that service isn't active anymore.
This only leaves GNU on the page after the move to GitLab, which is
kind of weird being a GNU project. We could of course start thanking
GitLab (and their CI service), but that list would still be short, so
I think removing is the better solution.