Unify dynamic relocation for Python scripts
Link out/share/lilypond/current/python to python/ in the source
directory and use it to find custom modules. This makes both
 $ ./out/bin/lilypond-book
and
 $ ./scripts/out/lilypond-book
work for in-tree and out-of-tree builds. The logic gives priority
to modules in the source directory (see #6015).

Moreover, also enable dynamic relocation for installation trees
with a default layout (version number instead of 'current'). This
is handy for 'make install DESTDIR=...'.