Fix iteration in Tie_performer::acknowledge_audio_element
If you compile LilyPond with -D_GLIBCXX_DEBUG, it aborts during "make
test" in this function, because it is invalid to increment an iterator
which is already at the end.

After erasing an element from the list, just return, instead of setting a
found flag which will exit the loop.