Ensure string for ps-quote in dump-pdf-bookmarks
As of 88fced680060f5187c06d6285a1ed061518fbd7e 'outline-bookmarks
is now always #t. Though, the relevant entry of the toc-alist may
not be a string, but a formatted markup.
Applies markup->string to fix this situation.
Test-case
  \markuplist \table-of-contents
  { \tocItem \markup \underline "xy" R1 }
Extends input/regression/toc-structered.ly to cover it.
Comment about issue 5999 no longer applies, as it is fixed now.
Removing.