Improve the part combiner's silence analysis
See the improvement in regtest part-combine-mmrest-shared.ly. A rest in
a measure by itself (because a whole-measure rest in the other part is
properly omitted) was previously vertically shifted, but now it is not.

This result is now consistent with part-combine-silence-mixed.ly, which
has the same combination of rest and whole-measure rest, but in the
first measure.