Fix draw-thick-zero-fret in fret-diagrams.scm
In 2.20.0 the thickness of lowest string was not scaled. Thus a visible
gap occured with increased size.  Evan bigger with additional increased
string-thickness-factor.  The regtest fret-diagrams-string-thickness.ly
did not cover this case, thus it was not detected with
8441002b90 Major cleanup in fret-diagrams.scm
Which made things worse in this regard.

This patches uses scaled thickness to fix the problem and extends
input/regression/fret-diagrams-string-thickness.ly to cover it.