Music_iterator: get scoped music properties
Add a protected internal_get_property () method allowing a music
iterator to look up a property of its own music or enclosing music.

This supports a planned \fine command that will work differently inside
a folded \repeat than outside.  The existing \unfolded feature is not
sufficient for this because \unfolded takes effect when the immediately
enclosing repeat is unfolded, but \fine needs to remain dormant until
all enclosing repeats are unfolded.