new command line options `-dpng-width` and `-dpng-height`
The idea is to let LilyPond create PNG images (from EPS files) with a given
pixel width and/or height.  If one of the dimensions is missing, the bbox of
the EPS file gets used to derive it from the other dimensions.

Note that ghostscript (version <= 9.52) has a bug: It creates empty PNG
images if the PNG height is larger than the PNG width.

  https://bugs.ghostscript.com/show_bug.cgi?id=702607

* scm/lily.scm (scheme-options-definitions): New options `png-width` and
  `png-height`.

* scm/backend-library.scm (gs-cmd-args): Add parameter `fit-page` to
  activate the `-dEPSFitPage` option of gs if necessary.
  Update all callers.
  (postscript->png): Add parameter `bbox` for adjusting the PNG width and/or
  height if the other dimension is missing.
  Also pass options `png-width` and `png-height` to `make-ps-images`.
  Update all callers.

* scm/framework-ps.scm: Updated.

* scm/ps-to-png.scm (make-ps-images): Implement new options.

* scm/define-markup-command.scm (get-postscript-bbox): Define with
  `define-public` so that we can access it in `framework-ps.scm`.