web: Remove links to missing binary packages
Since 2.21.0, we've dropped the following architectures:
 * linux-ppc
 * freebsd-x86, freebsd-64
 * darwin-ppc

I argue that they're equally meaningless for the current stable
release (2.20.0), so drop as well. If a user wants to download
the packages, they can be found at http://lilypond.org/downloads/.

While at it, clean up the version headers:
 * Download always refers to the current stable version.
 * Development always links to the current unstable binaries, but
   the manuals of the current commit.
This needs some love from the translators for the headers and warnings
in the respective download.itexi.