convertrules: Fix bad escapes
Single backslashes in arguments to re.sub() -- after expansion --
escape the next character, but only few characters are allowed.

For example, consider
    re.sub('\\commandA', '\\commandB', '\\commandA')
or the equivalent with raw strings:
    re.sub(r'\commandA', r'\commandB', r'\commandA')

Python 3 throws the following error for the first argument:
    re.error: bad escape \c at position 0
and, after fixing that, for the second:
    KeyError: '\\c'
    re.error: bad escape \c at position 0
(with a different backtrace).

The correct versions of the above call are:
    re.sub('\\\\commandA', '\\\\commandB', '\\commandA')
    re.sub(r'\\commandA', r'\\commandB', r'\commandA')
(or any mixture of this)