convertrules: Reduce quotation mark escaping
Escaping \' and \" in a string literal is always ok, but in many
cases not needed. Furthermore, r'\"' in a regular pattern seems
to work (as opposed to the replacement, see issue #6024), but is
deprecated -- even if not generating a DeprecationWarning nor an
error (only for unknown escapes of ASCII letters, since Python 3.6).