Fix similar backslashes in convert rules
As in the convert rule for \tocItem, there are many other lines
with missing r-prefixes or backslashes which are now fixed.

The file convertrules.py makes very inconsistent use of raw
strings, string quoting and escaping. This commit only fixes
the errors a consistent style would be a separate task.