help2man: Don't use locale
A completely installed `help2man` program comes with a bunch of locales,
which get used even while calling LilyPond's private version of the script.

However, LilyPond isn't set up (yet) to create non-English man pages; for
this reason, suppress the use of a locale to get identical builds that don't
depend on the build environment.