doc: Fix xrefs in translated manuals
The xref-maps for the translated manuals should depend on the master
map from the English version. This didn't work before because pattern
rules cannot be used to add additional dependencies. The results were
nondeterministic due to the number of jobs, evaluation order, and
dependency resolution.
In certain cases, this actually led to more correct xrefs in the
manuals: Due to an error in the script, the xref-maps contained
wrong entries for unnumbered nodes (anchor instead of filename).
Fix that by using the right entry of the tuple.