Turn on almost all Python warnings in build system
This option is recommended for testing as it enables warning developers
about pending deprecations before things break in newer versions of
Python. In particular, it uncovers any invalid escape sequences in
strings literals.

However, we silence the hundreds of ResourceWarnings emitted during
`make` due to files not being closed timely (it is generally
recommended not to rely on garbage collection for closing files).
Fixing this issue is out of the scope of this commit.