book_texinfo.py: Correctly handle commas in `@lydoctitle`
Commas in texinfo macro or command arguments should be protected since `,`
is the argument separator.  Note that for single-argument macros this is not
necessary; texinfo contains special code to handle that case.  However, file
`en/snippets.tely` redefines `@lydoctitle` to also call the multi-argument
command `@node`, which fails with unprotected commas.