Revert "lilypond-book: invoke process_cmd once per document"
This reverts commit a7514ab6b20236eef5de4b7cda2a44077e6da3ea.

Reason:

The texinfo snippet support decides to emit @include based on
existence of a .texidoc file. The .texidoc file only gets created by
the lilypond run, so moving the lilypond run to the end breaks the
snippet docs.

This is tricky to diagnose, because it doesn't happen in incremental
builds, and the find() that inspects the include path can also mask
this problem.