Revise handling of "no music in score" scenarios
ly:interpret-music-expression is used for partCombine music, quoted
music, and perhaps other things.  It does not raise a "no music in
score" warning and it creates contexts whether or not there is music.
This is the way it worked some time ago.

ly:run-translator is used for scores.  It raises a "skipping zero-
duration score" warning and does not create contexts (other than Global)
when either there is no music or the music has no duration.  The latter
condition is new.