Revert "Define notehead attachment points separately"
There are two problems: First, encodingdefs.ps wasn't updated leading
to breakage with -dmusic-font-encodings (implied by -OTeX-GS). Second,
the example for palm mute in fretted-strings.itely still references
the glyph "noteheads.u2do". This leads to a warning when building the
snippet and the glyph is just missing from the output.

This reverts commit 2d460deb7c33a962de8eaa591b2ee5800264892c.