Add skip-of-moment-span Scheme function
This function creates skip music fitting between a given start moment
and end moment.  This is useful to create skips with initial grace time.
The existing skip-of-length function does not serve because grace time
is not counted in music length.

Use the new function to fix an inconsistency in the timing of grace-only
percent repeats.