Do not check debugging switch DEBUG_SLUR_SCORING
The switch was introduced in
2ed0044bc7425734a1e44356b9afd68d46bb72e7
and was set to true by default in
`lily/include/main.hh` since then.

The code paths incorporating the debugging
functionality can be toggled with the
layout variable `debug-slur-scoring`,
therefore the preprocessor check is not
necessary.

Some regression tests make use of this
functionality, e.g. `slur-scoring.ly`.