Do not check debugging switch DEBUG_BEAM_SCORING
The switch was introduced in
849ed54c1a272dac222d975e3c8d3e532aaa8f27
and was set to true by default in
`lily/include/main.hh` since then.

The code paths incorporating the debugging
functionality can be toggled with the
layout variable `debug-beam-scoring`,
therefore the preprocessor check is not
necessary.

Some regression tests make use of this functionality,
e.g. `beam-flat-retain-direction.ly`.