Add \volta i,j,k function to mark volta-specific music
The \volta function marks a music expression as being limited to the
specified volte when the immediately enclosing repeat is unfolded.

\volta #'() has the natural effect of removing the music entirely when
it is unfolded.  Its complement is the \unfolded function, which masks
music until it is unfolded.

The user may supply \volta on the elements of \alternative {}.  If the
user does not, LilyPond adds it internally to match the legacy behavior
of repeating the first alternative the necessary number of times to
reach the repeat count.

The \volta function wraps music in VoltaSpeccedMusic.
Volta_specced_music_iterator sends VoltaSpanEvents announcing where
these spans of music start and stop (with the assigned labels).

Volta_engraver automatically constructs bracket labels based on
VoltaSpanEvents.  Automatic labels are coded to have lower precedence
than labels set in the repeatCommands context property.  I have chosen
not to regression-test interaction between automatic and manual labels
thoroughly because the NR says that repeatCommands is for unusual cases
and "may produce unexpected behavior;" however, I have added one
important case involving a volta bracket overhanging the next section.

Volta number formatting changes:
  1. Do not use a dash in a range spanning only two numbers.
  2. Use an em dash instead of two hyphens.

Volta_repeat_iterator ...

* no longer sets the volta label in the repeatCommands context property.
  Volta_specced_music_iterator and Volta_engraver have taken over.

* now ends the final alternative with an end-repeat bar if it applies to
  any volta other than the final volta

Repeat_acknowledge_engraver listens for volta span events (from \volta)
and inserts invisible bars when necessary to allow aligning brackets.

KNOWN ISSUE

The process of unfolding repeats filters volta-specific music but does
not recompute the length of the music.  Therefore, \volta can be used in
the main body of a repeated section to specify simultaneous alternatives
of the same length, but it can not generally be used to insert or remove
arbitrary music.