invoke-mf2pt1.sh: Avoid using python
The script called python from PATH which leads to errors, for example
on Ubuntu 20.04 without the package python-is-python3. While realpath
is indeed hard in a portable script, we only need absolute paths and
not actually resolve all symlinks. This is a matter of testing if the
first character is a slash and prepending $PWD otherwise.

Closes #6059