Fix pinglog
Calling (fdes->outport 2) opens a new port to the same descriptor,
which will subsequently be replaced by calls to ly:stderr-redirect.
As a result, the output appears in the individual log files unless
also using a log-file for the whole run (which was opened a second
time in append mode).

The (moderate) complexity in this change comes from not being able
to use (dup->outport 2) which crashes on mingw. Instead operate on
raw file descriptors and temporarily redirect the stderr for the
single "Processing ..." line. The notice about failed files is
redundant and will be fully restored for all cases in a subsequent
commit that outputs the final status to the original error stream.