Set up fallback fonts for Emmentaler.
FontConfig's standard configuration files use 'sans-serif' as the fallback
for all font families that haven't set up an explicit fallback.  However,
'sans-serif' is platform-dependent, which makes it hard to build
documentation with an exactly defined font setup.

This change fixes this; it uses the fonts for the 'LilyPond Serif' font
family alias.