Delete option-help.ly
This test prints the list of all options defined in scm/lily.scm,
including their help strings, with the values reflecting the current
invocation. Thus, the output changes when using a different number
of processes or passing the input files in a different order. This
complicates regression testing with varying numbers of job-count's.