lily/ttf.cc: Fix handling of Type 42 fonts
We split the `glyf` table into chunks aligned on glyph boundaries only if
there are short offsets.  The Type 42 specification expects that offsets
(and thus chunk lengths) are always even; this cannot be guaranteed for long
offsets, as counterexamples like the Korean Baekmuk fonts demonstrate.  Note
that this splitting is done only for compatibility with PS interpreters
earlier than version 2013, which was released in 1993...