Issue 6060: limit unproductive page spacing checks
In certain circumstances, where it is not possible to fulfill all spacing
requirements (e.g. making a score with only one note occupy two systems),
LilyPond is liable to spend a very long time checking layout with higher and
higher system counts - so long that it appears to have hung.

Add an additional check to the relevant loop that will break out when both of
the following are true:
- the results are infinitely bad
- increasing the system count has stopped working (i.e. incrementing sys_count
  doesn't actually result in there being more systems).

Closes #6060