SystemStartBracket: Take tips into account for bbox
Take the horizontal extent from the bracket line
to ensure correct positioning in X-direction (as before),
but take vertical extent from the whole bracket.

If a system starts with a bracket, e.g. in a
StaffGroup context, the bracket stencil is constructed
from a vertical line and the bracket tips.
The extent for the final stencil was taken only from
the bracket line, however. This can lead to incorrect
bounding boxes, which result in cropping when using the
EPS backend. This affects the '-dpreview' option or
the use of 'lilypond-book'.

This problem is hidden in many cases, because other
objects are higher or lower than the bracket tips so that
the incorrect bounding box will not show up.

Closes #720.