Exclude all ancestors of the merge base in print-gittxt.sh
The Dotted Range Notation "A..B" is equivalent to the range "^A B"
which includes all commits reachable from B, but excludes A and all
its ancestors. The script used $FP~1 which is the first parent and
aimed at including $FP without its ancestors.

There are two problems with this:
1. It requires that the first parent commit of the merge base is present
   in the local repository. This might not always be the case during CI
   which leads to issue #6061.
2. It does not properly handle merges and only excludes those commits
   reachable from the first parent of the merge commit.

Solve both issues by using $FP^! which is a shorthand to include $FP
but exclude all its parents. This notation was introduced in 2006 and
first appeared in git 1.4.4 which should be safe to assume.