Allow hairpin endpoints to be aligned to NoteColumn directions
By default hairpins extend to the extremes of the bound if there is
no adjacent hairpin or dynamic-text. This instead allows a hairpin to
extend to the LEFT, CENTER or RIGHT of NoteColumns by overriding the
property endpoint-alignments which is a pair of directions representing
the left and right ends of the hairpin.