System: Initialize all member variables
With Guile 2.2, it's possible to observe crashes like
    #0  System::derived_mark() const (this=0x55bd0b6a3730)
        at [...]/lily/include/smobs.hh:321
    #1  0x000055bd05f62f25 in Grob::mark_smob() const (this=0x55bd0b6a3730)
    #2  to #9 in libguile-2.2.so.1 and libgc.so.1
    #10 0x00007f5ed5bfe875 in scm_i_new_smob ()
    #11 0x000055bd05f5b00b in scm_new_smob (data=94270412975808, tc=13943)
    #12 Smob_base<Grob_array>::register_ptr(Grob_array*) (p=0x55bd0a7a6ec0)
    #13 Simple_smob<Grob_array>::smobbed_copy() const (this=<synthetic pointer>)
    #14 Grob_array::make_array()
    #15 0x000055bd060ccbdf in System::init_elements() (this=0x55bd0b6a3730)
    #16 0x000055bd060553f9 in Score_engraver::initialize() (this=0x55bd08b51ed0)

The reason is that System::System (SCM) did not initialize the pscore_
field and relied on Paper_score::typeset () to fill in the data. When
a GC happens between the two steps, System::derived_mark () calls
self_scm () (at lily/include/smobs.hh:321) on an uninitialized pointer.