Type-specific C++ grob acknowledgers
This change allows declaring a C++ grob acknowledger method as requiring
an Item or Spanner.  The type check is built into the callback
infrastructure so the acknowledger does not have to repeat the check.

In this change, only acknowledgers of spanner-interface and
item-interface use this new feature.

Changing other cases requires more thought.  If foo-interface implies
item-interface, then acknowledge_foo (Grob_info_t<Item>) is desirable;
ditto for spanners.  In other cases, dynamic casting is still required,
but in no case is it acceptable to dereference the result of a
dynamic_cast without first checking that it is non-null.