Bar numbers should not change with visibility
Move the bookkeeping for alternative bar numbering from
Bar_number_engraver to Timing_translator.  Closes #6068.

Also, the Bar_number_engraver now requires an actual bar line to trigger
creation of a bar number.  It used to create a bar number at the start
of a measure even if the Bar_engraver were removed.