Improve left-edge spacing of magnified staves
The Break_align_engraver automatically creates a LeftEdge grob to
participate in horizontal spacing calculations.

It had been creating the LeftEdge on handling the first break-alignable
(e.g. BarLine) or break-aligned (e.g. Clef) grob in a timestep, and
creating the LeftEdge as if it originated in the same context as the
grob triggering its creation.  The problem with that was that the
engravers for the break-alignable grobs that are currently handled
operate in Score context, so the LeftEdge would be created with the
space-alist values of Score context rather than the magnified values of
Staff context.

The improvement in this commit is to have Break_align_engraver defer
creating a LeftEdge until it is handling another break-aligned grob,
which is likely to originate in Staff or Voice.

This does not cure problems that \magnifyStaff causes in multi-staff
scores, but it closes #6084.