Reduce dynamic_cast in ancient ligature engravers
The ligature engravers assumed in various places that grobs received in
Ligature_engraver::acknowledge_ligature_head () are Items.  This change
uses Grob_info_t<Item> there and downstream, and removes casts that are
no longer necessary.

The only grob defined in scm/define-grobs.scm that has
ligature-head-interface is NoteHead, an Item, and it seems unlikely that
anyone would want to add ligature-head-interface to a non-Item any time
soon.