Paper_column_engraver: logic clean-up
Since fe9242659305dce587bd1fcdcc7b0ac62df25ad6,
start_translation_timestep () does not run in the same timestep as
initialize (), so there is no need for the first_ flag anymore.

made_columns_ was almost tautological with command_column_ != nullptr,
and "almost" seemed like an error, so get rid of made_columns_.

It is not possible for command_column_ to be null in finalize () because
it is set in initialize () and never nullified, so use it without fear.